requirements.txt
# core Langchain ecosystem
langchain==0.3.27
langchain-core
langchain-community
langchain-openai
langchain-text-splitters
langgraph
langchain-cohere

# supporting libs
uvicorn[standard]
fastapi
faiss-cpu
python-dotenv
ftfy
rank-bm25
nltk
opensearch-py
url-normalize

# downloading data and artifacts
huggingface-hub[hf_transfer]

# aws integration
boto3
requests-aws4auth

# monitoring
langsmith
